backend/requirements.txt
flask
python-dotenv
mysql-connector-python
uvicorn[standard]
pydantic
yfinance
apscheduler
flask_cors